Fixes errors with finding executable on windows
